Upgrade NUnit from v3.0.1 to v3.11.0
... because the project no longer builds with VisualStudio Community 7.6.11

```
DependencyGraph/DependencyGraphTest/ReflectorTest.cs(6,6): Error CS0619: 'TestFixtureAttribute' is obsolete: 'The NUnit framework shipped with Mono is deprecated and will be removed in a future release. It was based on NUnit 2.4 which is long outdated. Please move to the NUnit NuGet package or some other form of acquiring NUnit.' (CS0619) (DependencyGraphTest)
```